Subtitle: add vpts for agmp player [1/1]
PD#SWPL-134438

Problem:
Subtitle add vpts

Solution:
Subtitle add vpts

Verify:
local

+++ client/SubtitleClientLinux.cpp
@@ -212,7 +212,7 @@ int SubtitleClientLinux::resetForSeek() {
 
 }
 
-int SubtitleClientLinux::updateVideoPos(int32_t pos) {
+int SubtitleClientLinux::updateVideoPos(int64_t pos) {
     int ret = SendMethodCall(
             format_string("subtitle.ctrl.%d.%d.%d", SUBTITLE_UPDATEVIDEOPTS, mSessionId,
                           pos));
